refactor: extract ExportHistorySummaryView.swift
Extracts the ExportHistorySummaryView View (~45 lines) from
TranscriptRecoveryAndHistoryViews.swift into its own file.

Byte-identical move of lines 86-130. Only imports SwiftUI.

TranscriptRecoveryAndHistoryViews.swift: 130 → 84 lines.
Tests: 667.

⚠️ Manual smoke check recommended: verify export history summary
renders in the transcript view.
